removed channel_i constraining--fixing tricky bug
channel constraining is a BAD idea....ie: when the user tries to set Ch8
even though they set numChannels to only 7, it is BAD to constrain their
channel input and set what they think is their Ch8 value on Ch7...as
this is a really hard to find bug. Rather, it is best to not do
anything, thereby letting the user know when they see stuff not working
that they are not actually setting the channel. So...I fixed this in the
whole .cpp file...NO MORE CONSTRAINING CHANNEL INDEXES INPUT BY THE
USER!!! Just don't do anything if their channel_i value is invalid.
DONE! This stinking bug cost me a couple hrs lost time.